APIBackend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.13
django-celery-results
django-cors-headers
djangorestframework
djongo==1.3.6
dnspython==2.6.1
h11==0.14.0
idna==3.8
joblib==1.4.2
kombu==5.4.0
lxml==5.3.0
mongoengine==0.28.2
nltk==3.9.1
ntscraper==0.3.16
outcome==1.3.0.post0
prompt_toolkit==3.0.47
pymongo==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
regex==2024.7.24
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.2.4
textblob==0.18.0.post0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
